[STYLE] #110 separate point section components
